clean up / refactor `runner.py` helpers
