+++ src/groundhog_hpc/runner.py
@@ -1,13 +1,11 @@
 import warnings
-from hashlib import sha1
-from pathlib import Path
 from typing import Callable
 from uuid import UUID
 
 from groundhog_hpc.errors import RemoteExecutionError
 from groundhog_hpc.serialization import deserialize, serialize
 from groundhog_hpc.settings import DEFAULT_USER_CONFIG
-from groundhog_hpc.utils import get_groundhog_version_spec
+from groundhog_hpc.templating import template_shell_command
 
 warnings.filterwarnings(
     "ignore",
@@ -15,63 +13,36 @@ warnings.filterwarnings(
     module="globus_compute_sdk",
 )
 
-SHELL_COMMAND_TEMPLATE = """
-cat > {script_name}.py << 'EOF'
-{script_contents}
-EOF
-cat > {script_name}.in << 'END'
-{payload}
-END
-$(python -c 'import uv; print(uv.find_uv_bin())') run -qq --managed-python --with {version_spec} \\
-  {script_name}.py {function_name} {script_name}.in > {script_name}.stdout \\
-  && cat {script_name}.out
-"""
-# note: working directory is ~/.globus_compute/uep.<endpoint uuids>/tasks_working_dir
-
 
 def script_to_callable(
-    user_script: str,
+    script_path: str,
     function_name: str,
     endpoint: str,
     walltime: int | None = None,
     user_endpoint_config: dict | None = None,
-    script_path: str | None = None,
 ) -> Callable:
     """Create callable corresponding to the named function from a user's script.
 
     The created function accepts the same arguments as the original named function, but
-    dispatches to a shell function on the remote endpoint.
-
-    NOTE: The function must expect json-serializable input and return json-serializable output.
+    dispatches to a shell function on the target remote endpoint.
     """
     import globus_compute_sdk as gc  # lazy import so cryptography bindings don't break remote endpoint
 
     config = DEFAULT_USER_CONFIG.copy()
     config.update(user_endpoint_config or {})
 
-    script_hash = _script_hash_prefix(user_script)
-    script_basename = (
-        _extract_script_basename(script_path) if script_path else "groundhog"
+    shell_command = template_shell_command(script_path, function_name)
+    shell_function = gc.ShellFunction(
+        shell_command, walltime=walltime, name=function_name
     )
-    script_name = f"{script_basename}-{script_hash}"
-    script_contents = _inject_script_boilerplate(
-        user_script, function_name, script_name
-    )
-
-    version_spec = get_groundhog_version_spec()
 
     def run(*args, **kwargs):
-        shell_fn = gc.ShellFunction(cmd=SHELL_COMMAND_TEMPLATE, walltime=walltime)
         payload = serialize((args, kwargs))
 
         with gc.Executor(UUID(endpoint), user_endpoint_config=config) as executor:
             future = executor.submit(
-                shell_fn,
-                script_name=script_name,
-                script_contents=script_contents,
-                function_name=function_name,
+                shell_function,
                 payload=payload,
-                version_spec=version_spec,
             )
 
             shell_result: gc.ShellResult = future.result()
@@ -88,38 +59,23 @@ def script_to_callable(
     return run
 
 
-def _script_hash_prefix(contents: str, length=8) -> str:
-    return str(sha1(bytes(contents, "utf-8")).hexdigest()[:length])
+def pre_register_shell_function(
+    script_path: str, function_name: str, walltime: int | None = None
+) -> UUID:
+    """Pre-register a `ShellFunction` corresponding to the named function in a
+    script and return its function UUID.
 
+    Note that the registered function will expect a single `payload` kwarg which
+    should be a serialized str, and will return a serialized str to be
+    deserialized.
+    """
+    import globus_compute_sdk as gc
 
-def _extract_script_basename(script_path: str) -> str:
-    return Path(script_path).stem
-
-
-def _inject_script_boilerplate(
-    user_script: str,
-    function_name: str,
-    script_name: str,
-) -> str:
-    assert "__main__" not in user_script, (
-        "invalid user script: can't define custom `__main__` logic"
+    client = gc.Client()
+    shell_command = template_shell_command(script_path, function_name)
+    shell_function = gc.ShellFunction(
+        shell_command, walltime=walltime, name=function_name
     )
-    payload_path = f"{script_name}.in"
-    outfile_path = f"{script_name}.out"
 
-    script = f"""{user_script}
-if __name__ == "__main__":
-    import sys
-
-    from groundhog_hpc.serialization import serialize, deserialize
-
-    with open('{payload_path}', 'r') as f_in:
-        payload = f_in.read()
-        args, kwargs = deserialize(payload)
-
-    results = {function_name}(*args, **kwargs)
-    with open('{outfile_path}', 'w+') as f_out:
-        contents = serialize(results)
-        f_out.write(contents)
-"""
-    return script
+    function_id = client.register_function(shell_function, public=True)
+    return function_id
